requirements.txt
playwright>=1.40.0
pdfplumber>=0.10.0
python-docx>=1.0.0
openpyxl>=3.1.0
pyyaml>=6.0
